CRM_project/requirement.txt
django
djongo==1.3.6
djangorestframework
pymongo==3.12.2
